requirements.txt
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.0
matplotlib==3.6.2
seaborn==0.12.1
tensorflow==2.5.0
pillow==8.2.0